+++ src/Device/kernels/ReSTIR/SPMIS/ComputeOffsets.h
@@ -11,10 +11,10 @@
 
 #ifdef __KERNELCC__
 GLOBAL_KERNEL_SIGNATURE(void)
-ReSTIR_SPMIS_ComputeOffsets(unsigned int* cell_counters, unsigned int* global_cell_offset_counter, unsigned int* cell_offsets, unsigned int size)
+ReSTIR_SPMIS_ComputeOffsets(unsigned short int* cell_counters, unsigned int* global_cell_offset_counter, unsigned int* cell_offsets, unsigned int size)
 #else
 GLOBAL_KERNEL_SIGNATURE(void)
-inline ReSTIR_SPMIS_ComputeOffsets(AtomicType<unsigned int>* cell_counters,
+inline ReSTIR_SPMIS_ComputeOffsets(AtomicType<unsigned short int>* cell_counters,
 								   AtomicType<unsigned int>* global_cell_offset_counter,
 								   unsigned int* cell_offsets,
 								   unsigned int size,
@@ -28,9 +28,9 @@ inline ReSTIR_SPMIS_ComputeOffsets(AtomicType<unsigned int>* cell_counters,
 	if (x >= size)
 		return;
 
-	unsigned int hash_cell_index = x;
-	unsigned int cell_count		 = hippt::atomic_load(&cell_counters[hash_cell_index]);
-	unsigned int cell_offset	 = hippt::atomic_fetch_add(global_cell_offset_counter, cell_count);
+	unsigned int hash_cell_index  = x;
+	unsigned short int cell_count = hippt::atomic_load(&cell_counters[hash_cell_index]);
+	unsigned int cell_offset	  = hippt::atomic_fetch_add(global_cell_offset_counter, (unsigned int)cell_count);
 
 	cell_offsets[hash_cell_index] = cell_offset;
 }
